# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CodeMind AI API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/organizations/{org_id}/github/repositories": {
      "get": {
        "tags": [
          "github"
        ],
        "summary": "List Available Repositories",
        "operationId": "list_available_repositories_api_organizations__org_id__github_repositories_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepositoryListingResponse"
                  },
                  "title": "Response List Available Repositories Api Organizations  Org Id  Github Repositories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{org_id}/repositories": {
      "post": {
        "tags": [
          "repositories"
        ],
        "summary": "Add Repository",
        "operationId": "add_repository_api_organizations__org_id__repositories_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRepositoryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "List Repositories",
        "operationId": "list_repositories_api_organizations__org_id__repositories_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepositoryResponse"
                  },
                  "title": "Response List Repositories Api Organizations  Org Id  Repositories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{org_id}/repositories/{repo_id}": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "Get Repository",
        "operationId": "get_repository_api_organizations__org_id__repositories__repo_id__get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          },
          {
            "name": "repo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddRepositoryRequest": {
        "properties": {
          "external_repo_id": {
            "type": "string",
            "title": "External Repo Id"
          }
        },
        "type": "object",
        "required": [
          "external_repo_id"
        ],
        "title": "AddRepositoryRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RepositoryListingResponse": {
        "properties": {
          "external_repo_id": {
            "type": "string",
            "title": "External Repo Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch"
          }
        },
        "type": "object",
        "required": [
          "external_repo_id",
          "full_name",
          "default_branch"
        ],
        "title": "RepositoryListingResponse"
      },
      "RepositoryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch"
          },
          "latest_index_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Index Status"
          },
          "latest_index_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Index Id"
          },
          "latest_analysis_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Analysis Status"
          },
          "latest_analysis_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Analysis Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "default_branch"
        ],
        "title": "RepositoryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}